DIV-4303 - renamed test files to be consistent and swapped actual/expect
