+++ src/test/java/uk/gov/hmcts/reform/divorce/caseformatterservice/strategy/reasonfordivorce/DesertionStrategyUTest.java
@@ -1,6 +1,8 @@
 package uk.gov.hmcts.reform.divorce.caseformatterservice.strategy.reasonfordivorce;
 
+import org.junit.Before;
 import org.junit.Test;
+import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.ccd.CoreCaseData;
 import uk.gov.hmcts.reform.divorce.caseformatterservice.domain.model.usersession.DivorceSession;
 
 import java.text.ParseException;
@@ -9,11 +11,17 @@ import java.util.Locale;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsEqual.equalTo;
+import static org.junit.Assert.assertEquals;
 
 public class DesertionStrategyUTest {
     private static final String DESERTION = "desertion";
 
-    private final DesertionStrategy desertionStrategy = new DesertionStrategy();
+    private DesertionStrategy desertionStrategy;
+
+    @Before
+    public void setup() {
+        desertionStrategy = new DesertionStrategy();
+    }
 
     @Test
     public void testDesertionStatementOfCase() throws ParseException {
@@ -40,5 +48,44 @@ public class DesertionStrategyUTest {
 
         assertThat(derivedStatementOfCase, equalTo(
             "I have been deserted by my null on the 01 January 1970.\n"));
+    }
+
+    @Test
+    public void testSetLivedApartFieldsFromDivorceSessionSetsFieldsOnCoreCaseDataObject() {
+        //given
+        DivorceSession divorceSession = new DivorceSession();
+        divorceSession.setLivedApartEntireTime("Yes");
+        divorceSession.setLivedTogetherMoreTimeThanPermitted("No");
+        divorceSession.setTimeLivedTogetherPermitted("1 month");
+
+        CoreCaseData coreCaseData = new CoreCaseData();
+
+        //when
+        desertionStrategy.setLivedApartFieldsFromDivorceSession(divorceSession, coreCaseData);
+
+        //then
+        assertEquals("YES", coreCaseData.getDesertionLivedApartEntireTime());
+        assertEquals("NO", coreCaseData.getDesertionLivedTogetherMoreTimeThanPermitted());
+        assertEquals("1 month", coreCaseData.getDesertionTimeTogetherPermitted());
+    }
+
+    @Test
+    public void testLivedApartFieldsFromCoreCaseDataSetsFieldsoOnDivorceSessionObject() {
+        //given
+        CoreCaseData coreCaseData = new CoreCaseData();
+
+        coreCaseData.setDesertionLivedApartEntireTime("YES");
+        coreCaseData.setDesertionLivedTogetherMoreTimeThanPermitted("NO");
+        coreCaseData.setDesertionTimeTogetherPermitted("1 month");
+
+        DivorceSession divorceSession = new DivorceSession();
+
+        //when
+        desertionStrategy.setLivedApartFieldsFromCoreCaseData(coreCaseData, divorceSession);
+
+        //then
+        assertEquals("Yes", divorceSession.getLivedApartEntireTime());
+        assertEquals("No", divorceSession.getLivedTogetherMoreTimeThanPermitted());
+        assertEquals("1 month", divorceSession.getTimeLivedTogetherPermitted());
     }
 }
